added key check when uploading to s3

+++ aws-glue/upload_to_s3.py
@@ -8,7 +8,7 @@ try:
 except Exception as e:
     print('Error - libraries: {}'.format(e))
 
-FILE_PATH = "D:/Study Materials/My git projects/Aws-prac/"
+FILE_PATH = "D:/Study Materials/My git projects/Aws/"
 BUCKET_NAME = 'junn-test-glue'
 
 s3_client = boto3.client('s3')
@@ -20,19 +20,36 @@ def get_files():
     except Exception as e:
         print("Error - get_files: {}".format(e))
 
+def is_key_exists(curr_key):
+    '''
+    Checks if the file is already stored
+    '''
+    try:
+        s3_client.get_object(Bucket=BUCKET_NAME,
+                             Key= curr_key)
+        return True
+    except s3_client.exceptions.NoSuchKey:
+        print("ERROR: is_key_exists - {}".format(e))
+        return False
+
 def upload_to_bucket():
     try:
-       # To do: function to check if the file is stored already
         all_file_names = get_files()
         for file in all_file_names:
             obj_name= str(os.path.basename(file))
-            print("uploading file - {}".format(obj_name))
-            response= s3_client.upload_file(
-                                FILE_PATH+'/'+obj_name,
-                                Bucket=BUCKET_NAME,
-                                Key=obj_name
-        )   
-            print("uploaded file - {}".format(obj_name))
+            if not is_key_exists(obj_name):
+                
+                print("uploading file - {}".format(obj_name))
+
+                response= s3_client.upload_file(
+                                    FILE_PATH+'/'+obj_name,
+                                    Bucket=BUCKET_NAME,
+                                    Key=obj_name
+                                    )   
+                print("uploaded file - {}".format(obj_name))
+            else:
+                print("File already exists")
+
     except ClientError as e:
         logging.error(e)
         return False
